refactor: move out accordion logic into isolated component
